# swagger.yaml
openapi: 3.0.0
info:
    title: JOBS API PROD
    contact: {}
    version: '1.0'
servers:
    - url: https://jobs-api-node-94.herokuapp.com/api/v1
      variables: {}
paths:
...
    /job/{id}:
        parameters:
            - in: path
              name: id
              schema:
                  type: string
              required: true
              description: The job ID
        get:
            tags:
                - Job
            summary: GET JOB
            operationId: GETJOB
            parameters: []
            responses:
                '200':
                    description: ''
                    headers: {}
            deprecated: false
        patch:
            tags:
                - Job
            summary: UPDATE JOB
            operationId: UPDATEJOB
            parameters: []
            requestBody:
                description: ''
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UPDATEJOBREQUEST'
                        example:
                            company: Space
                            position: Front end dev 2
                required: true
            responses:
                '200':
                    description: ''
                    headers: {}
            deprecated: false
        delete:
            tags:
                - Job
            summary: DELETE JOB
            operationId: DELETEJOB
            parameters: []
            responses:
                '200':
                    description: ''
                    headers: {}
            deprecated: false
components:
    schemas:
...
        UPDATEJOBREQUEST:
            title: UPDATEJOBREQUEST
            required:
                - company
                - position
            type: object
            properties:
                company:
                    type: string
                position:
                    type: string
            example:
                company: Space
                position: Front end dev 2
    securitySchemes:
        httpBearer:
            type: http
            scheme: bearer
security:
    - httpBearer: []
...